+++ database/migrations/2022_07_29_082549_create_product_sizes_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
         Schema::create('product_sizes', function (Blueprint $table) {
             $table->id();
 
-            $table->integer('product_id');
-            $table->integer('size_id');
+            $table->unsignedBigInteger('product_id');
+            $table->unsignedBigInteger('size_id');
 
             $table->index('product_id', 'product_sizes_product_idx');
             $table->foreign('product_id', 'product_sizes_product_fk')->on('products')->references('id');
